Improve soft shadows using pcf and poisson sampling

+++ package/shaders/shadowmap-frag.glsl
@@ -1,11 +1,10 @@
 #version 300 es
 
-precision mediump float;
+precision highp float;
 
-// Ouput data
 layout(location = 0) out float fragmentdepth;
 
-void main(){
-    // Not really needed, OpenGL does it anyway
+void main()
+{
     fragmentdepth = gl_FragCoord.z;
 }
